refactor: 클래스를 역할에 따라 작게 분리하고 패키지를 변경하라
