+++ src/main/java/lotto/config/AppConfig.java
@@ -1,5 +1,6 @@
 package lotto.config;
 
+import lotto.controller.ApplicationController;
 import lotto.service.LottoService;
 import lotto.view.InputValidator;
 import lotto.view.InputView;
@@ -7,8 +8,8 @@ import lotto.controller.LottoController;
 import lotto.view.OutputView;
 
 public class AppConfig {
-    public AppRunner createAppRunner() {
-        return new AppRunner(createInputView(), createInputValidator(), createOutputView(), createLottoController());
+    public ApplicationController createAppRunner() {
+        return new ApplicationController(createInputView(), createInputValidator(), createOutputView(), createLottoController());
     }
 
     private InputView createInputView() {
