+++ src/main/java/lotto/view/InputValidator.java
@@ -1,7 +1,7 @@
 package lotto.view;
 
-import static lotto.MessageContainer.NEITHER_DIGIT_NOR_DELIMITER_ERROR;
-import static lotto.MessageContainer.NON_DIGIT_ERROR;
+import static lotto.view.ViewMessageContainer.NEITHER_DIGIT_NOR_DELIMITER_ERROR;
+import static lotto.view.ViewMessageContainer.NON_DIGIT_ERROR;
 import static lotto.view.ViewConstants.DIGIT_AND_DELIMITER_ONLY_PATTERN;
 import static lotto.view.ViewConstants.DIGIT_ONLY_PATTERN;
